Move subscription request to requests.py
Remove requests module for simplicity. If it becomes large we can always change it later on.
