+++ chargebee_byte/client.py
@@ -1,6 +1,6 @@
 import requests
 
-from chargebee_byte.requests.subscription_request import SubscriptionRequest
+from chargebee_byte.requests import SubscriptionRequest
 
 
 class Client(object):
